src/core/config/database_setup.py
import subprocess
from alembic import command
from alembic.config import Config
from psycopg2.extras import RealDictCursor
from pymongo import MongoClient
from pymongo.errors import ConnectionFailure
from sqlalchemy import create_engine, text, MetaData, Table, Column, Integer, String, Float, DateTime, JSON, Boolean, Text
from sqlalchemy.dialects.postgresql import UUID
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
from sqlalchemy.orm import sessionmaker, declarative_base
import alembic
import asyncpg
import psycopg2
import redis.asyncio as redis
from __future__ import annotations
from dataclasses import dataclass
from datetime import datetime
from pathlib import Path
from typing import Any, Dict, List, Optional
import asyncio
import json
import logging
import os
import sys

#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
파일명: database_setup.py
모듈: 데이터베이스 초기 설정 및 스키마 생성
목적: PostgreSQL, Redis, MongoDB 등 데이터베이스 초기화

Author: AI Assistant
Created: 2025-01-27
Modified: 2025-01-27
Version: 1.0.0

Dependencies:
    - Python 3.11+
    - sqlalchemy, asyncpg (PostgreSQL)
    - redis (Redis)
    - pymongo (MongoDB)
    - alembic (마이그레이션)
"""



try:
    PSYCOPG2_AVAILABLE = True
except ImportError:
    PSYCOPG2_AVAILABLE = False

try:
    ASYNCPG_AVAILABLE = True
except ImportError:
    ASYNCPG_AVAILABLE = False

try:
    SQLALCHEMY_AVAILABLE = True
except ImportError:
    SQLALCHEMY_AVAILABLE = False

try:
    REDIS_AVAILABLE = True
except ImportError:
    REDIS_AVAILABLE = False

try:
    PYMONGO_AVAILABLE = True
except ImportError:
    PYMONGO_AVAILABLE = False

try:
    ALEMBIC_AVAILABLE = True
except ImportError:
    ALEMBIC_AVAILABLE = False

# 로깅 설정
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('database_setup.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

# SQLAlchemy Base
if SQLALCHEMY_AVAILABLE:
    Base = declarative_base()
else:
    Base = None


@dataclass
class DatabaseConfig:
    """데이터베이스 설정"""
    # PostgreSQL 설정
    postgres_host: str = "localhost"
    postgres_port: int = 5432
    postgres_db: str = "trading_data"
    postgres_user: str = "postgres"
    postgres_password: str = "password"

    # Redis 설정
    redis_host: str = "localhost"
    redis_port: int = 6379
    redis_db: int = 0
    redis_password: Optional[str] = None

    # MongoDB 설정
    mongodb_host: str = "localhost"
    mongodb_port: int = 27017
    mongodb_db: str = "trading_data"
    mongodb_user: Optional[str] = None
    mongodb_password: Optional[str] = None

    # 설정
    create_tables: bool = True
    create_indexes: bool = True
    create_views: bool = True
    backup_existing: bool = True

    def __post_init__(self) -> None:
        # 환경변수에서 설정 로드
        self.postgres_host = os.getenv('POSTGRES_HOST', self.postgres_host)
        self.postgres_port = int(os.getenv('POSTGRES_PORT', str(self.postgres_port)))
        self.postgres_db = os.getenv('POSTGRES_DB', self.postgres_db)
        self.postgres_user = os.getenv('POSTGRES_USER', self.postgres_user)
        self.postgres_password = os.getenv('POSTGRES_PASSWORD', self.postgres_password)

        self.redis_host = os.getenv('REDIS_HOST', self.redis_host)
        self.redis_port = int(os.getenv('REDIS_PORT', str(self.redis_port)))
        self.redis_db = int(os.getenv('REDIS_DB', str(self.redis_db)))
        self.redis_password = os.getenv('REDIS_PASSWORD', self.redis_password)

        self.mongodb_host = os.getenv('MONGODB_HOST', self.mongodb_host)
        self.mongodb_port = int(os.getenv('MONGODB_PORT', str(self.mongodb_port)))
        self.mongodb_db = os.getenv('MONGODB_DB', self.mongodb_db)
        self.mongodb_user = os.getenv('MONGODB_USER', self.mongodb_user)
        self.mongodb_password = os.getenv('MONGODB_PASSWORD', self.mongodb_password)


class DatabaseSetup:
    # ... (rest of the code)
