src/core/config/qubole_setup.py
from qds_sdk.commands import HiveCommand
from qds_sdk.qubole import Qubole
import boto3
from __future__ import annotations
from typing import Any
import Dict
import List, Optional, Tuple
import asyncio
import json
import logging
import os

#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
파일명: qubole_setup.py
모듈: Qubole 클라우드 환경 설정
목적: Qubole Data Lake 초기 설정 및 테이블 생성

Author: AI Assistant
Created: 2025-01-27
Modified: 2025-01-27
Version: 1.0.0

Dependencies:
    - Python 3.11+
    - qds-sdk (Qubole Data Service SDK)
    - boto3 (AWS S3)
    - asyncio, logging, os, json
"""



try:
    BOTO3_AVAILABLE = True
except ImportError:
    BOTO3_AVAILABLE = False

try:
    QUBOLE_AVAILABLE = True
except ImportError:
    QUBOLE_AVAILABLE = False

# 로깅 설정
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class QuboleSetup:
    """Qubole 설정 클래스"""

    def __init__(self) -> None:
        # Qubole API 설정
        self.api_token: Optional[str] = os.getenv('QUBOLE_API_TOKEN')
        self.api_url: str = os.getenv('QUBOLE_API_URL', 'https://api.qubole.com')

        # AWS S3 설정
        self.s3_bucket: str = os.getenv('S3_BUCKET', 'trading-data-lake')
        self.s3_region: str = os.getenv('S3_REGION', 'us-east-1')
        self.aws_access_key: Optional[str] = os.getenv('AWS_ACCESS_KEY_ID')
        self.aws_secret_key: Optional[str] = os.getenv('AWS_SECRET_ACCESS_KEY')

        # 데이터 레이크 경로
        self.data_lake_path: str = f"s3://{self.s3_bucket}"
        self.raw_data_path: str = f"s3://{self.s3_bucket}/raw"
        self.processed_data_path: str = f"s3://{self.s3_bucket}/processed"
        self.ml_data_path: str = f"s3://{self.s3_bucket}/ml"

    async def setup_qubole(self) -> None:
        """Qubole 초기 설정"""
        logger.info("Qubole 초기 설정 시작")

        try:
            # 1. Qubole 클라이언트 초기화
            await self._initialize_qubole()

            # 2. S3 버킷 설정
            await self._setup_s3_bucket()

            # 3. Hive 테이블 생성
            await self._create_hive_tables()

            # 4. 데이터 레이크 구조 생성
            await self._create_data_lake_structure()

            # 5. 연결 테스트
            await self._test_connections()

            logger.info("Qubole 초기 설정 완료")

        except Exception as e:
            logger.error(f"Qubole 초기 설정 실패: {e}")
            raise

    # ... (나머지 함수들)
