src/core/config/requirements.txt
# Trading AI System - 통합 의존성 파일
# Python 3.11+ 필요

# =============================================================================
# 핵심 의존성
# =============================================================================

# 데이터 검증 및 설정
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
dynaconf==3.2.4

# 비동기 처리
aiohttp==3.9.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1
async-timeout==4.0.0

# 데이터 처리
pandas==2.1.4
numpy==1.24.0
scikit-learn==1.3.0
openpyxl==3.1.0

# 데이터베이스
sqlalchemy==2.0.0
asyncpg==0.30.0
redis==5.0.0
aioredis==2.0.1
psycopg2-binary==2.9.0

# =============================================================================
# 금융 데이터 수집
# =============================================================================

# DART API
OpenDartReader==0.1.5
dart-fss==0.3.0
xmltodict==0.13.0

# 한투 KIS API
pykis==1.0.0

# 기타 금융 데이터
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10

# =============================================================================
# 기술적 분석
# =============================================================================

# 기술적 지표
ta==0.10.2
ta-lib==0.4.28
talib-binary==0.4.26
pandas-ta==0.3.14b0

# 시계열 분석
statsmodels==0.14.0
arch==6.2.0

# =============================================================================
# 머신러닝 및 AI
# =============================================================================

# 기본 ML
xgboost==2.0.2
tensorflow==2.15.0
torch==2.1.2

# 고급 AI
transformers==4.36.2
google-generativeai==0.3.2

# 최적화
scipy==1.11.4
cvxpy==1.4.1

# =============================================================================
# 감정 분석
# =============================================================================

textblob==0.17.1
vaderSentiment==3.3.2
konlpy==0.6.0

# =============================================================================
# 시각화
# =============================================================================

matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# =============================================================================
# 웹 프레임워크 및 API
# =============================================================================

fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
streamlit==1.28.0
dash==2.14.0

# =============================================================================
# 실시간 처리
# =============================================================================

celery==5.3.4
kafka-python==2.0.2
orderbook==0.1.0
market-microstructure==0.2.0

# =============================================================================
# 모니터링 및 로깅
# =============================================================================

psutil==5.9.0
prometheus-client==0.19.0
sentry-sdk==1.38.0
python-json-logger==2.0.7
structlog==23.2.0

# =============================================================================
# 스케줄링
# =============================================================================

schedule==1.2.0
apscheduler==3.10.1

# =============================================================================
# 캐싱 및 저장
# =============================================================================

diskcache==5.6.3
cachetools==5.3.2
boto3==1.34.0
botocore==1.34.0

# 시계열 데이터베이스 (선택사항)
influxdb-client==1.38.0

# =============================================================================
# 개발 도구
# =============================================================================

# 테스트
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.11

# 코드 품질
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.13.2
pre-commit==3.6.0
bandit==1.7.5

# 문서화
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# 개발 환경
jupyter==1.0.0
ipython==8.17.2

# =============================================================================
# 유틸리티
# =============================================================================

click==8.1.7
rich==13.7.0
tqdm==4.66.0
python-dateutil==2.8.2
pytz==2023.3
python-telegram-bot==20.7

# 성능 최적화
numba==0.58.0
cython==3.0.0

# 보안
cryptography==41.0.0

# Firebase (푸시 알림)
firebase-admin==6.4.0

# =============================================================================
# 선택적 의존성 (필요시 주석 해제)
# =============================================================================

# 메모리 프로파일링
# memory-profiler==0.61.0

# 고급 모니터링
# datadog==0.44.0

# GPU 가속 (CUDA 지원)
# cupy-cuda11x==12.0.0

# 분산 처리
# dask==2023.11.0
# ray==2.8.0 